+++ tests/Reader/FallbackReaderTestCase.php
@@ -9,25 +9,19 @@
 
 declare(strict_types=1);
 
-namespace Spiral\Tests\Attributes\Reader\Complex;
+namespace Spiral\Tests\Attributes\Reader;
 
-use Spiral\Attributes\AnnotationReader;
-use Spiral\Attributes\AttributeReader;
-use Spiral\Attributes\Composite\SelectiveReader;
+use Spiral\Attributes\Internal\FallbackAttributeReader;
 use Spiral\Attributes\ReaderInterface;
 
 /**
  * @group unit
  * @group reader
- * @group complex
  */
-class SelectiveReaderTest extends ComplexTestCase
+class FallbackReaderTestCase extends ComplexTestCase
 {
     protected function getReader(): ReaderInterface
     {
-        return new SelectiveReader([
-            new AttributeReader(),
-            new AnnotationReader()
-        ]);
+        return new FallbackAttributeReader();
     }
 }
