+++ tests/Reader/MergeReaderTestCase.php
@@ -9,7 +9,7 @@
 
 declare(strict_types=1);
 
-namespace Spiral\Tests\Attributes\Reader\Complex;
+namespace Spiral\Tests\Attributes\Reader;
 
 use Spiral\Attributes\AnnotationReader;
 use Spiral\Attributes\AttributeReader;
@@ -18,13 +18,12 @@ use Spiral\Attributes\ReaderInterface;
 
 /**
  * Count of annotations should be x2 of original values, except non-supported
- * by Doctrine reader ({@see DoctrineReaderTest} for more information).
+ * by Doctrine reader ({@see DoctrineReaderTestCase} for more information).
  *
  * @group unit
  * @group reader
- * @group complex
  */
-class MergeReaderTest extends ComplexTestCase
+class MergeReaderTestCase extends ComplexTestCase
 {
     protected $classMetadataCount = 2;
     protected $methodMetadataCount = 2;
